refactor(core): Use consistent CSRF state validation across oauth controllers
